Fixed a bug where the timer would only reset back to the original duration once
Changed the way the checkGuess function worked so that the program doesn't have to check the type of the exercise passed in, as this is already being performed in getNextExercise(). Instead, I created a new state variable, correctAnswer, and set this inside of getNextExercise().

Updated the UI to hide the Submit button after the timer has run out

Updated the UI to have better feedback than "Yippee!" and ":(". This includes displaying the correct answer to the user in the event that they get the answer wrong.

Updated the UI to display the amount of CORRECT guesses out of the total exercises played as opposed to the current number of exercises played